Move AwardedUser to new folder
