+++ app/components/Payments/styledComponents.js
@@ -1,6 +1,5 @@
 import styled from 'styled-components';
 import ToggleButtonGroup from '@material-ui/lab/ToggleButtonGroup';
-import { Link } from 'react-router-dom';
 
 import {
   ErrorSuccessBanner,
@@ -14,9 +13,7 @@ import {
   defaultFontSize,
   dividerBorder,
   fundingText,
-  hoverLinkColor,
   lightBlueColor,
-  rewardColor,
   successGreen,
 } from 'defaultStyleHelper';
 import { mediaQueriesByDevice } from 'utils/breakpoints';
@@ -27,13 +24,6 @@ export const Amount = styled.div`
   align-self: center;
   font-size: 2.4rem;
   font-weight: 500;
-`;
-
-export const AwardedUserWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  font-size: 1.4rem;
-  padding: 0 0.5rem 0.5rem;
 `;
 
 export const BalanceAmount = styled.h3`
@@ -77,18 +67,6 @@ export const ChargeValue = styled.div`
   text-align: end;
 `;
 
-export const CoinWrapper = styled.div`
-  align-items: center;
-  display: flex;
-  justify-content: center;
-
-  svg {
-    height: 2.25rem;
-    padding-right: 0.5rem;
-    width: 2.25rem;
-  }
-`;
-
 export const ConfirmAmount = styled.div`
   color: ${({ isNegative }) => (isNegative ? buttonRed : 'inherit')};
   white-space: nowrap;
@@ -111,12 +89,6 @@ export const ConfirmWrapper = styled.div`
   font-weight: ${({ isBold }) => (isBold ? '400' : '600')};
   justify-content: space-between;
   padding: 1rem;
-`;
-
-export const ContentWrapper = styled.div`
-  align-items: center;
-  align-self: center;
-  display: flex;
 `;
 
 export const CreditCardViewContainer = styled.div`
@@ -162,16 +134,6 @@ export const InputWrapper = styled.div`
   width: auto;
 `;
 
-export const LinkWrapper = styled.div`
-  display: flex;
-  margin: 1rem;
-
-  svg {
-    height: 2rem;
-    width: 2rem;
-  }
-`;
-
 export const PaymentInformationWrapper = styled.div`
   display: flex;
   flex-direction: column;
@@ -193,33 +155,8 @@ export const OverviewWrapper = styled.div`
   padding: 3rem;
 `;
 
-export const Rewarded = styled.span`
-  align-self: center;
-  color: ${rewardColor};
-  font-size: 1.4rem;
-  font-weight: 500;
-  padding: 0.5rem 0.5rem 2rem;
-`;
-
 export const StyledErrorSuccessBanner = styled(ErrorSuccessBanner)`
   margin: 1rem;
-`;
-
-export const StyledExternalLink = styled.a`
-  padding-left: 0.5rem;
-  white-space: nowrap;
-
-  &:hover {
-    color: ${hoverLinkColor};
-  }
-`;
-
-export const StyledInternalLink = styled(Link)`
-  font-weight: 500;
-
-  &:hover {
-    color: ${hoverLinkColor};
-  }
 `;
 
 export const StyledLabel = styled.div`
